Moves conversions for ParamVals Moves teh conversions for ParamVals from the conversion mod to the attr_map mod
