+++ attribution/src/attr_map.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
+use std::convert::TryInto;
 use syn::parse::Parse;
 use syn::parse::ParseStream;
-use syn::parse::Result;
 use syn::Lit;
 use syn::Meta;
 use syn::NestedMeta;
@@ -33,7 +33,7 @@ impl Default for Parameters {
 }
 
 impl Parse for Parameters {
-    fn parse(buffer: ParseStream) -> Result<Self> {
+    fn parse(buffer: ParseStream) -> syn::parse::Result<Self> {
         let mut attribute_map = Self::new();
 
         while !buffer.is_empty() {
@@ -105,6 +105,60 @@ impl From<Lit> for ParamVal {
     }
 }
 
+#[derive(Debug, PartialEq)]
+pub struct TryIntoParamValError;
+
+impl TryInto<bool> for ParamVal {
+    type Error = TryIntoParamValError;
+    fn try_into(self) -> Result<bool, Self::Error> {
+        if let ParamVal::Bool(b) = self {
+            Ok(b)
+        } else {
+            Err(TryIntoParamValError {})
+        }
+    }
+}
+
+impl TryInto<u64> for ParamVal {
+    type Error = TryIntoParamValError;
+    fn try_into(self) -> Result<u64, Self::Error> {
+        if let ParamVal::Int(i) = self {
+            Ok(i)
+        } else {
+            Err(TryIntoParamValError {})
+        }
+    }
+}
+
+impl TryInto<String> for ParamVal {
+    type Error = TryIntoParamValError;
+    fn try_into(self) -> Result<String, Self::Error> {
+        if let ParamVal::Str(s) = self {
+            Ok(s)
+        } else {
+            Err(TryIntoParamValError {})
+        }
+    }
+}
+
+impl From<&str> for ParamVal {
+    fn from(src: &str) -> Self {
+        ParamVal::Str(src.into())
+    }
+}
+
+impl From<u64> for ParamVal {
+    fn from(src: u64) -> Self {
+        ParamVal::Int(src)
+    }
+}
+
+impl From<bool> for ParamVal {
+    fn from(src: bool) -> Self {
+        ParamVal::Bool(src)
+    }
+}
+
 #[cfg(test)]
 mod attr_val_tests {
 
@@ -143,5 +197,26 @@ mod attr_val_tests {
                 case.2
             );
         }
+    }
+
+    #[test]
+    fn bool_conversion() {
+        let left = ParamVal::Bool(true).try_into();
+        let right = Ok(true);
+        assert_eq!(left, right)
+    }
+
+    #[test]
+    fn int_conversion() {
+        let left = ParamVal::Int(1).try_into();
+        let right = Ok(1);
+        assert_eq!(left, right)
+    }
+
+    #[test]
+    fn str_conversion() {
+        let left: Result<String, TryIntoParamValError> = ParamVal::Str("hello".into()).try_into();
+        let right: Result<String, TryIntoParamValError> = Ok("hello".into());
+        assert_eq!(left, right)
     }
 }
